ZENKO-2518 webpack cleanup
