+++ .gitignore
@@ -1,4 +1,6 @@
 /public/assets/js
+/public/assets/fonts
+/public/assets/index.html
 
 yarn-error.log
