+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "test:lint": "eslint --ext js --ext jsx src",
-    "start": "webpack-dev-server --mode development",
-    "build": "webpack --mode production"
+    "start:dev": "webpack-dev-server --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js"
   },
   "keywords": [],
   "author": "",
@@ -32,12 +32,14 @@
     "flow-bin": "^0.118.0",
     "flow-type": "^1.0.1",
     "flow-typed": "^2.6.2",
-    "html-loader": "^0.5.5",
+    "html-loader": "^1.0.0",
+    "html-webpack-plugin": "^4.0.3",
     "redux-logger": "^3.0.6",
     "style-loader": "^1.1.3",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.3"
+    "webpack-dev-server": "^3.10.3",
+    "webpack-merge": "^4.2.2"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "5.7.2",
