+++ public/assets/config.json
@@ -1,0 +1,3 @@
+{
+    "api_endpoint": "http://127.0.0.1:5000"
+}
