+++ src/index-template.html
@@ -6,7 +6,5 @@
 </head>
 <body>
   <div id="app"></div>
-
-  <script src="http://localhost:8383/zenko_ui.js"></script>
 </body>
 </html>
