+++ webpack.common.js
@@ -1,0 +1,72 @@
+const path = require('path');
+const HtmlWebPackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+    entry: {
+        zenko_ui: './src/react/App',
+    },
+    output: {
+        filename: 'js/[name].js',
+        path: path.resolve(__dirname, 'public/assets'),
+        publicPath: '/',
+    },
+    resolve: {
+        modules: ['node_modules'],
+        extensions: [
+            '.js',
+            '.jsx',
+        ],
+    },
+    module: {
+        rules: [
+            {
+                test: /\.(js|jsx)$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                },
+            },
+            {
+                test: /\.html$/,
+                use: [{
+                    loader: 'html-loader',
+                }],
+            },
+            {
+                test: /\.jpe?g$|\.gif$|\.png$|\.ttf$|\.eot$|\.svg$/,
+                use: [
+                    {
+                        loader: 'file-loader',
+                        options: {
+                            name: '[name].[ext]?[hash]',
+                            outputPath: 'fonts/',
+                        },
+                    },
+                ],
+            },
+            {
+                test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+                use: [
+                    {
+                        loader: 'url-loader',
+                        options: {
+                            limit: 10000,
+                            mimetype: 'application/fontwoff',
+                            outputPath: 'fonts/',
+                        },
+                    },
+                ],
+            },
+            {
+                test: /\.css$/,
+                loader: 'style-loader!css-loader',
+            },
+        ],
+    },
+    plugins: [
+        new HtmlWebPackPlugin({
+            template: './src/index-template.html',
+            filename: './index.html',
+        }),
+    ],
+};
