+++ webpack.prod.js
@@ -1,0 +1,12 @@
+const webpack = require('webpack');
+const merge = require('webpack-merge');
+const common = require('./webpack.common');
+
+module.exports = merge(common, {
+    devtool: 'hidden-source-map',
+    plugins: [
+        new webpack.DefinePlugin({
+            'process.env.NODE_ENV': JSON.stringify('production'),
+        }),
+    ],
+});
